Ignore combo in sequece of keys
For example, if we have I + P combo and try to type "vip", the combo is
triggered because i and p are pressed in short duration. To avoid such errors,
check the previous key action is enough old. If v is pressed within 500ms
before i, vip will be recognized as "vip", not as "v + key combo".